refactor: group app hook owner folders
