+++ src/hooks/app-bootstrap/useAppBootstrapEffects.ts
@@ -1,13 +1,13 @@
 import { useEffect } from 'react';
 import type { Dispatch, SetStateAction } from 'react';
 
-import type { FestivalItem, Place } from '../types/core';
-import type { SessionUser } from '../types/auth';
-import type { StampState } from '../types/review';
-import type { MyPageResponse } from '../types/my-page';
-import { bootstrapFestivalLoader } from './app-bootstrap/bootstrapFestivalLoader';
-import { bootstrapMapSession } from './app-bootstrap/bootstrapMapSession';
-import { clearAuthQueryParams } from './useAppRouteState';
+import type { FestivalItem, Place } from '../../types/core';
+import type { SessionUser } from '../../types/auth';
+import type { StampState } from '../../types/review';
+import type { MyPageResponse } from '../../types/my-page';
+import { bootstrapFestivalLoader } from './bootstrapFestivalLoader';
+import { bootstrapMapSession } from './bootstrapMapSession';
+import { clearAuthQueryParams } from '../app-route/useAppRouteState';
 import type { AppBootstrapSharedRefs } from './useAppBootstrapSharedRefs';
 
 type SetProviders = (providers: Array<{ key: 'naver' | 'kakao'; label: string; isEnabled: boolean; loginUrl: string | null }>) => void;
