+++ src/hooks/app-coordinator/buildAppShellCoordinatorResult.ts
