+++ src/hooks/app-coordinator/useAppShellCoordinator.types.ts
@@ -1,15 +1,15 @@
 import type {
   getInitialMapViewport,
   useAppRouteState,
-} from './useAppRouteState';
-import type { useAuthDomainState } from './useAuthDomainState';
-import type { useAppShellRuntimeState } from './useAppShellRuntimeState';
-import type { useAppPageRuntimeState } from './useAppPageRuntimeState';
-import type { useAppDataState } from './useAppDataState';
-import type { useMapDomainState } from './useMapDomainState';
-import type { useMyPageDomainState } from './useMyPageDomainState';
-import type { useReturnViewDomainState } from './useReturnViewDomainState';
-import type { useReviewDomainState } from './useReviewDomainState';
+} from '../app-route/useAppRouteState';
+import type { useAuthDomainState } from '../useAuthDomainState';
+import type { useAppShellRuntimeState } from '../useAppShellRuntimeState';
+import type { useAppPageRuntimeState } from '../useAppPageRuntimeState';
+import type { useAppDataState } from '../useAppDataState';
+import type { useMapDomainState } from '../useMapDomainState';
+import type { useMyPageDomainState } from '../useMyPageDomainState';
+import type { useReturnViewDomainState } from '../useReturnViewDomainState';
+import type { useReviewDomainState } from '../useReviewDomainState';
 
 export type RouteState = ReturnType<typeof useAppRouteState>;
 export type DomainState = {
