+++ src/hooks/app-tab-loaders/useAppTabWarmup.ts
@@ -1,8 +1,8 @@
 import { useEffect } from 'react';
-import type { Tab } from '../types/core';
-import type { SessionUser } from '../types/auth';
-import type { MyPageResponse } from '../types/my-page';
-import type { AdminSummaryResponse } from '../types/admin';
+import type { Tab } from '../../types/core';
+import type { SessionUser } from '../../types/auth';
+import type { MyPageResponse } from '../../types/my-page';
+import type { AdminSummaryResponse } from '../../types/admin';
 
 interface UseAppTabWarmupParams {
   activeTab: Tab;
